+++ actions/bind_test.go
@@ -1,14 +1,9 @@
 package actions
 
 import (
-	"bytes"
-	"net/http"
-	"net/http/httptest"
 	"reflect"
 	"strings"
 	"testing"
-
-	internalactions "github.com/nlstn/go-odata/internal/actions"
 )
 
 type discountInput struct {
@@ -84,69 +79,36 @@ func TestBindParams_TypeMismatch(t *testing.T) {
 	}
 }
 
-func TestBindParams_UsesExistingBinding(t *testing.T) {
+func TestBindParams_RepeatCallsUseConsistentBinding(t *testing.T) {
 	params := map[string]interface{}{
-		boundStructKey: &discountInput{Percentage: 55},
+		"percentage": 55.0,
 	}
 
-	result, err := BindParams[discountInput](params)
+	first, err := BindParams[discountInput](params)
 	if err != nil {
-		t.Fatalf("BindParams() unexpected error: %v", err)
+		t.Fatalf("BindParams() first call unexpected error: %v", err)
 	}
-	if result.Percentage != 55 {
-		t.Fatalf("BindParams() percentage = %v, want 55", result.Percentage)
+	if first.Percentage != 55 {
+		t.Fatalf("BindParams() first call percentage = %v, want 55", first.Percentage)
+	}
+
+	second, err := BindParams[discountInput](params)
+	if err != nil {
+		t.Fatalf("BindParams() second call unexpected error: %v", err)
+	}
+	if second.Percentage != 55 {
+		t.Fatalf("BindParams() second call percentage = %v, want 55", second.Percentage)
+	}
+	if !reflect.DeepEqual(first, second) {
+		t.Fatalf("BindParams() first and second results differ: first=%#v, second=%#v", first, second)
 	}
 
 	ptrResult, err := BindParams[*discountInput](params)
 	if err != nil {
-		t.Fatalf("BindParams() pointer unexpected error: %v", err)
+		t.Fatalf("BindParams() pointer call unexpected error: %v", err)
 	}
 	if ptrResult == nil || ptrResult.Percentage != 55 {
-		t.Fatalf("BindParams() pointer = %#v, want percentage 55", ptrResult)
+		t.Fatalf("BindParams() pointer result = %#v, want percentage 55", ptrResult)
 	}
 }
 
-func TestBindParams_WithStructBinding(t *testing.T) {
-	type actionInput struct {
-		Name  string  `mapstructure:"name"`
-		Count int64   `mapstructure:"count"`
-		Note  *string `mapstructure:"note,omitempty"`
-	}
-
-	body := `{"name":"Widget","count":5}`
-	req := httptest.NewRequest(http.MethodPost, "/Apply", bytes.NewBufferString(body))
-	req.Header.Set("Content-Type", "application/json")
-
-	defs, err := internalactions.ParameterDefinitionsFromStruct(reflect.TypeOf(actionInput{}))
-	if err != nil {
-		t.Fatalf("ParameterDefinitionsFromStruct() unexpected error: %v", err)
-	}
-
-	params, err := internalactions.ParseActionParameters(req, defs, reflect.TypeOf(actionInput{}))
-	if err != nil {
-		t.Fatalf("ParseActionParameters() unexpected error: %v", err)
-	}
-
-	bound, err := BindParams[actionInput](params)
-	if err != nil {
-		t.Fatalf("BindParams() unexpected error: %v", err)
-	}
-
-	if bound.Name != "Widget" {
-		t.Fatalf("BindParams() name = %q, want %q", bound.Name, "Widget")
-	}
-	if bound.Count != 5 {
-		t.Fatalf("BindParams() count = %d, want 5", bound.Count)
-	}
-	if bound.Note != nil {
-		t.Fatalf("BindParams() expected nil note, got %#v", bound.Note)
-	}
-
-	ptrResult, err := BindParams[*actionInput](params)
-	if err != nil {
-		t.Fatalf("BindParams() pointer unexpected error: %v", err)
-	}
-	if ptrResult == nil || ptrResult.Count != 5 {
-		t.Fatalf("BindParams() pointer = %#v, want count 5", ptrResult)
-	}
-}
